sample_gcp/requirements.txt
flask==3.1.0
gunicorn
numpy==1.26.4
pillow==11.0.0
tensorflow==2.17.1